Rename schedule-next-step.ts to scheduling.ts
Summary:
This increases consistency of our agent naming.

Test Plan:
pnpm lint
